# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: EV Charging Booking API
  description: |
    API lengkap untuk sistem Booking Charger Kendaraan Listrik.
    
    ## Autentikasi
    Gunakan Bearer token JWT di header Authorization:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Base URL
    - Development: `http://localhost:3000/api`
    - Production: `https://api.ev-charging.id/api`
    
    ## Rate Limiting
    - Auth endpoints: 5 req/menit per IP
    - Semua lainnya: 100 req/menit per user
    
  version: "1.0.0"
  contact:
    name: Tim EV Charging
    email: dev@ev-charging.id
servers:
  - url: http://localhost:3000/api
    description: Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
      required: [error, message]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "usr_01H..."
        email:
          type: string
          format: email
          example: "budi@email.com"
        name:
          type: string
          example: "Budi Santoso"
        phone:
          type: string
          example: "+628123456789"
        ev_plate:
          type: string
          example: "B 1234 EV"
        role:
          type: string
          enum: [USER, OPERATOR, ADMIN]
        created_at:
          type: string
          format: date-time
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Daftar akun baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
                  minLength: 2
                phone:
                  type: string
                  example: "+628123456789"
                ev_plate:
                  type: string
                  example: "B 1234 EV"
      responses:
        "201":
          description: Akun berhasil dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  access_token:
                    type: string
                  expires_in:
                    type: integer
                    description: Detik hingga access token expired
        "400":
          description: Validasi input gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Email sudah terdaftar
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login dan dapatkan JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login berhasil
          headers:
            Set-Cookie:
              description: HttpOnly cookie berisi refresh token
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  access_token:
                    type: string
                  expires_in:
                    type: integer
        "401":
          description: Email atau password salah
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token menggunakan refresh token dari cookie
      responses:
        "200":
          description: Token berhasil di-refresh
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  expires_in:
                    type: integer
        "401":
          description: Refresh token tidak valid atau expired
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout dan invalidate refresh token
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logout berhasil

  # USERS ===========================================================================
  /users/me:
    get:
      tags: [Users]
      summary: Dapatkan profil saya
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [Users]
      summary: Update profil saya
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                ev_plate:
                  type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # STATIONS ===========================================================================
...